코드 정리 및 auth 기능 구조 생성
