+++ Back_End/src/functions/signJWT.js
@@ -1,14 +1,12 @@
-"use strict";
-
 const jwt = require('jsonwebtoken');
 const config = require('config');
 const ACCESS_SECRET_KEY = config.get('JWT.access_secret_key');
 const REFRESH_SECRET_KEY = config.get('JWT.refresh_secret_key');
 
-const sign_jwt = {
+const signJWT = {
     access(payload) {
         return jwt.sign(payload, ACCESS_SECRET_KEY, {
-            expiresIn: '15m', 
+            expiresIn: '15m',
             issuer: config.get('JWT.issuer'),
         });
     },
@@ -32,6 +30,6 @@ const sign_jwt = {
             }
         );
     }
-}
+};
 
-module.exports = sign_jwt;
+module.exports = signJWT;
