+++ Back_End/src/middleware/verifyJWT.js
@@ -1,17 +1,12 @@
-"use strict";
-
 const jwt = require('jsonwebtoken');
 const config = require('config');
 const ACCESS_SECRET_KEY = config.get('JWT.access_secret_key');
 
-exports.auth = (req, res, next) => {
-    // 인증 완료
+exports.verifyJWT = (req, res, next) => {
     try {
-        // 요청 헤더에 저장된 토큰(req.headers.authorization)과 비밀키를 사용하여 토큰을 req.decoded에 반환
         req.decoded = jwt.verify(req.headers.authorization, ACCESS_SECRET_KEY);
         return next();
     }
-    // 인증 실패
     catch (error) {
         // 유효시간이 초과된 경우
         if (error.name === "TokenExpiredError") {
